# scripts/docs/openapi/openapi-v1-20260315-163040.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "MambaSplit.Api",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/v1/groups":  {
                                         "post":  {
                                                      "tags":  [
                                                                   "Group"
                                                               ],
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/CreateGroupRequest"
                                                                                                                              }
                                                                                                               },
                                                                                          "text/json":  {
                                                                                                            "schema":  {
                                                                                                                           "$ref":  "#/components/schemas/CreateGroupRequest"
                                                                                                                       }
                                                                                                        },
                                                                                          "application/*+json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/CreateGroupRequest"
                                                                                                                                }
                                                                                                                 }
                                                                                      }
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK",
                                                                                    "content":  {
                                                                                                    "text/plain":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/GroupDto"
                                                                                                                                  }
                                                                                                                   },
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/GroupDto"
                                                                                                                                        }
                                                                                                                         },
                                                                                                    "text/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/GroupDto"
                                                                                                                                 }
                                                                                                                  }
                                                                                                }
                                                                                }
                                                                    }
                                                  },
                                         "get":  {
                                                     "tags":  [
                                                                  "Group"
                                                              ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "content":  {
                                                                                                   "text/plain":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "array",
                                                                                                                                     "items":  {
                                                                                                                                                   "$ref":  "#/components/schemas/GroupDto"
                                                                                                                                               }
                                                                                                                                 }
                                                                                                                  },
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "type":  "array",
                                                                                                                                           "items":  {
                                                                                                                                                         "$ref":  "#/components/schemas/GroupDto"
                                                                                                                                                     }
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "type":  "array",
                                                                                                                                    "items":  {
                                                                                                                                                  "$ref":  "#/components/schemas/GroupDto"
                                                                                                                                              }
                                                                                                                                }
                                                                                                                 }
                                                                                               }
                                                                               }
                                                                   }
                                                 }
                                     },
                  "/api/v1/groups/{groupId}/details":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Group"
                                                                                ],
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "groupId",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  {
                                                                                                                     "text/plain":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/GroupDetailsDto"
                                                                                                                                                   }
                                                                                                                                    },
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/GroupDetailsDto"
                                                                                                                                                         }
                                                                                                                                          },
                                                                                                                     "text/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/GroupDetailsDto"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
                  "/api/v1/groups/{groupId}":  {
                                                   "delete":  {
                                                                  "tags":  [
                                                                               "Group"
                                                                           ],
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "groupId",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "string"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "OK"
                                                                                            }
                                                                                }
                                                              }
                                               },
...
                  "/api/v1/groups/{groupId}/invites/{token}":  {
                                                                   "delete":  {
                                                                                  "tags":  [
                                                                                               "Group"
                                                                                           ],
                                                                                  "parameters":  [
                                                                                                     {
                                                                                                         "name":  "groupId",
                                                                                                         "in":  "path",
                                                                                                         "required":  true,
                                                                                                         "schema":  {
                                                                                                                        "type":  "string"
                                                                                                                    }
                                                                                                     },
                                                                                                     {
                                                                                                         "name":  "token",
                                                                                                         "in":  "path",
                                                                                                         "required":  true,
                                                                                                         "schema":  {
                                                                                                                        "type":  "string"
                                                                                                                    }
                                                                                                     }
                                                                                                 ],
                                                                                  "responses":  {
                                                                                                    "200":  {
                                                                                                                "description":  "OK"
                                                                                                            }
                                                                                                }
                                                                              }
                                                               },
                  "/api/v1/groups/{groupId}/invites/by-id/{inviteId}":  {
                                                                            "delete":  {
                                                                                           "tags":  [
                                                                                                        "Group"
                                                                                                    ],
                                                                                           "parameters":  [
                                                                                                              {
                                                                                                                  "name":  "groupId",
                                                                                                                  "in":  "path",
                                                                                                                  "required":  true,
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "string"
                                                                                                                             }
                                                                                                              },
                                                                                                              {
                                                                                                                  "name":  "inviteId",
                                                                                                                  "in":  "path",
                                                                                                                  "required":  true,
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "string"
                                                                                                                             }
                                                                                                              }
                                                                                                          ],
                                                                                           "responses":  {
                                                                                                             "200":  {
                                                                                                                         "description":  "OK"
                                                                                                                     }
                                                                                                         }
                                                                                       }
                                                                        },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateGroupRequest":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "name":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              }
                                                                                 },
                                                                  "additionalProperties":  false
                                                              },
...
                                       "GroupDto":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "string",
                                                                                      "nullable":  true
                                                                                  },
                                                                           "name":  {
                                                                                        "type":  "string",
                                                                                        "nullable":  true
                                                                                    }
                                                                       },
                                                        "additionalProperties":  false
                                                    },
...
                                   }
                   }
}